+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/ChoiceTypeTest.php
@@ -89,7 +89,7 @@ class ChoiceTypeTest extends TypeTestCase
     }
 
     /**
-     * expectedException \Symfony\Component\Form\Exception\FormException
+     * expectedException \Symfony\Component\Form\Exception\Exception
      */
     public function testEitherChoiceListOrChoicesMustBeSet()
     {